adjustements to damage and energy values

+++ Source/Descent/Public/PawnBase.h
@@ -30,9 +30,9 @@ public:
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Variables")
 	bool bHasKey = false;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Variables")
-	int EnergyCounter = 100;
+	float EnergyCounter = 100.0f;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Variables")
-	int ShieldCounter = 100;
+	float ShieldCounter = 100.0f;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Variables")
 	int HostagesCounter = 0;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Variables")
